+++ src/PageComposer/Livewire/Elements/Photo.php
@@ -18,8 +18,6 @@ class Photo extends Component
 
     public $photo, $elementId;
 
-    public $target;
-
     public $breakpoints = [
         'none',
         'sm',
@@ -57,8 +55,6 @@ class Photo extends Component
     public function updateData()
     {
         $this->showElementInputs = false;
-
-        $this->dispatch('elementUpdated.' . $this->target, data: $this->data, itemKey: $this->itemKey);
     }
 
     public function deleteExistingPhoto()