+++ src/resources/views/livewire/row-component.blade.php
@@ -93,7 +93,7 @@
     <div wire:sortable="updateColumnOrder" wire:sortable.options="{ animation: 100 }" class="flex space-x-4 justify-left transition pt-0 @if (count($row['columns']) > 0) pt-4 @endif">
         @foreach ($this->sortedColumns as $columnKey => $column)
             <div wire:sortable.item="{{ $columnKey }}" class="{{ $this->columnWidth($column['column_size']) }}">
-                <livewire:column-component :column="$column" :key="uniqid()" :columnKey="$columnKey" :previewMode="$previewMode" target="{{ $source }}" />
+                <livewire:column-component :column="$column" :key="uniqid()" :columnKey="$columnKey" :previewMode="$previewMode" />
             </div>
         @endforeach
     </div>